requirements.txt
gradio==4.13.0
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
sentencepiece==0.1.99
accelerate==0.25.0
langchain==0.0.267
faiss-cpu==1.7.4
sentence-transformers==2.7.0
pymupdf==1.23.7
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
nltk==3.8.1
spacy==3.6.1
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.3.0
pytesseract==0.3.10
Pillow==10.0.0
pdfplumber==0.10.2
tqdm==4.66.1
